feat: Setup Daily Sync
Have created functions to test whether it is time to sync with the servers. This could be changed to daily or at set intervals by changing when _ALREADY_SYNCED is set.  Have not set up the function to set when to sync. This could be a specific time or every couple hours.